backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlmodel==0.0.16
sqlalchemy==2.0.31
pydantic==2.8.2
python-multipart==0.0.9
Pillow==10.4.0
numpy==2.0.1
opencv-python==4.10.0.84
easyocr==1.7.1
pandas==2.2.2
jinja2==3.1.4
reportlab==4.2.2
matplotlib==3.9.0
seaborn==0.13.2
